Migrated string tests
